Split native secondary pane views
