+++ docs/DECISIONS.md
@@ -174,7 +174,7 @@
 - Settings surface contracts live in `QuillCodeSettingsSurface.swift`. `WorkspaceSettingsSurface`, `WorkspaceSettingsUpdate`, Computer Use requirement rows, TrustedRouter sign-in copy, Computer Use permission labels, and compatibility decoding stay beside the settings contract; `WorkspaceSurface.swift` should only carry the aggregate `settings` payload and pass current runtime state into it.
 - Terminal and browser utility panes live in `QuillCodeTerminalBrowserPaneView.swift`. Terminal history rows, browser navigation, page snapshots, comments, and pane-local draft state should evolve there; `WorkspaceSwiftUIView.swift` should only decide when the panes are visible and route the resulting actions.
 - Terminal surface contracts live in `QuillCodeTerminalSurface.swift`. `TerminalSurface` and `TerminalCommandSurface` own run/clear availability, cwd labels, command lifecycle labels, and execution-context propagation; `WorkspaceSurface.swift` should only carry the aggregate `terminal` payload and assemble it from terminal state.
-- Secondary utility panes live in `QuillCodeSecondaryPanesView.swift`. Extensions, Memories, Automations, their count/status pills, empty states, and compact cards should evolve there; `WorkspaceSwiftUIView.swift` should only decide pane placement and route commands.
+- Secondary utility pane chrome lives in `QuillCodeSecondaryPanesView.swift`. Extensions, Memories, and Automations each have focused native view files: `QuillCodeExtensionsPaneView.swift`, `QuillCodeMemoriesPaneView.swift`, and `QuillCodeAutomationsPaneView.swift`. Shared count/status pills and empty states should evolve in the chrome file, while pane-specific compact cards evolve beside their own command routing; `WorkspaceSwiftUIView.swift` should only decide pane placement and route commands.
 - Secondary pane surface contracts live in `QuillCodeSecondaryPaneSurface.swift`. `WorkspaceExtensionsSurface`, `WorkspaceMemoriesSurface`, `WorkspaceAutomationsSurface`, extension manifest rows, memory rows, and automation workflow rows should evolve beside the secondary-pane native and HTML renderers; `WorkspaceSurface.swift` should only carry the aggregate slots and assemble them from model state.
 - Workspace modal presentation lives in `QuillCodeWorkspaceSheets.swift`, while each dialog family has a named owner: `QuillCodeCommandPaletteDialog.swift` owns command palette rows and icon mapping, `QuillCodeSearchAndShortcutDialogs.swift` owns chat search and keyboard shortcuts, `QuillCodeWorktreeDialogs.swift` owns worktree create/remove drafts and sheets, `QuillCodeWorkspaceDialogs.swift` owns rename sheets, and `QuillCodeDialogChrome.swift` owns shared dialog header, section, empty-state, and labeled-field primitives. `WorkspaceSwiftUIView.swift` should only decide when sheets are presented and route completed actions.
 - Workspace view command routing is planned by `WorkspaceViewCommandPlanner`. The planner owns command-ID to view-action mapping, selected thread/project rename lookup, worktree sheet intents, and composer focus rules. `WorkspaceSwiftUIView.swift` should execute typed `WorkspaceViewCommandAction` values and avoid growing another command-ID switch as Codex-parity commands expand.
